backend/requirements.txt
fastapi
uvicorn
requests
soniox
python-dotenv
